Unify LegacyPythonCallbackParser with Parser abstract class
[ci skip-rust-tests]
