+++ src/python/pants/engine/internals/examples/parsers.py
@@ -37,7 +37,6 @@ def _import(typename):
 
 class JsonParser(Parser):
     def __init__(self, symbol_table):
-        super().__init__()
         self.symbol_table = symbol_table
 
     def _as_type(self, type_or_name):
@@ -212,7 +211,6 @@ class PythonAssignmentsParser(Parser):
     """
 
     def __init__(self, symbol_table):
-        super().__init__()
         self.symbol_table = symbol_table
 
     @memoized_property
@@ -265,7 +263,6 @@ class PythonCallbacksParser(Parser):
     """
 
     def __init__(self, symbol_table):
-        super().__init__()
         self.symbol_table = symbol_table
         self._lock = threading.Lock()
