+++ src/python/pants/init/BUILD
@@ -48,9 +48,10 @@ python_library(
     'src/python/pants/engine/internals:core_abstractions',
     'src/python/pants/engine/internals:graph',
     'src/python/pants/engine/internals:native',
+    'src/python/pants/engine/internals:mapper',
     'src/python/pants/engine/internals:options_parsing',
+    'src/python/pants/engine/internals:parser',
     'src/python/pants/engine/internals:scheduler',
-    'src/python/pants/engine/legacy:parser',
     'src/python/pants/engine/legacy:structs',
     'src/python/pants/engine:console',
     'src/python/pants/engine:fs',
